Moved EventProjection and GenerateGuid to common project
